+++ lib/capybara_accessible_selectors/filter_set.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 Capybara::Selector::FilterSet.add(:capybara_accessible_selectors)
+require "capybara_accessible_selectors/filters/current"
 require "capybara_accessible_selectors/filters/described_by"
 require "capybara_accessible_selectors/filters/fieldset"
 require "capybara_accessible_selectors/filters/focused"
@@ -15,8 +16,8 @@ require "capybara_accessible_selectors/filters/validation_error"
   field: %i[focused fieldset described_by validation_error],
   file_field: %i[focused fieldset described_by validation_error],
   fillable_field: %i[focused fieldset described_by validation_error],
-  link: %i[focused fieldset],
-  link_or_button: %i[focused fieldset],
+  link: %i[current focused fieldset],
+  link_or_button: %i[current focused fieldset],
   radio_button: %i[focused fieldset described_by validation_error],
   select: %i[focused fieldset described_by validation_error],
   xpath: %i[focused]