+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Add `:menu` and `:menuitem` selector
+- Add `current:` filter for `:link` and `:link_or_button` selectors
 
 ## v0.8.2